use upx

+++ build_extension.bat
@@ -1,4 +1,4 @@
-curl https://raw.githubusercontent.com/GM82Project/gm82gex/main/gm82gex.py -o gm82gex.py
+if not exist gm82gex.py curl https://raw.githubusercontent.com/GM82Project/gm82gex/main/gm82gex.py -o gm82gex.py
 if exist gm82live.dll gm82gex.py gm82live.gej
 
 pause
